Update Kotlin to 1.4.30 and remove android-ktx
* Use JDK 11 on CI
* Update kotlin to 1.4.30
* Bump build number
* Remove android-ktx and move it to androidMain source set
* Migrate Gradle scripts from Groovy to Kotlin DSL
