--- android-ktx/build.gradle
@@ -1,54 +1,0 @@
-plugins {
-    id 'com.android.library'
-    id 'kotlin-android'
-    id 'kotlin-platform-android'
-    id 'digital.wup.android-maven-publish' version '3.6.2'
-    id 'mirego.release' version '2.0'
-    id 'mirego.publish' version '1.0'
-}
-
-repositories {
-    google()
-    jcenter()
-    mavenLocal()
-    mavenCentral()
-    maven { url('https://s3.amazonaws.com/mirego-maven/public') }
-}
-
-group 'com.mirego.trikot.bluetooth'
-
-android {
-    compileSdkVersion 30
-
-    defaultConfig {
-        minSdkVersion 21
-        targetSdkVersion 30
-    }
-}
-
-configurations.all {
-    exclude group: "org.reactivestreams"
-}
-
-dependencies {
-    api 'com.mirego.trikot:bluetooth:1.0.1'
-    implementation "com.mirego.trikot:trikotFoundation:$trikot_foundation_version"
-    implementation "com.mirego.trikot:android-ktx:$trikot_streams_android_ktx_version"
-    implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
-    implementation 'androidx.lifecycle:lifecycle-reactivestreams-ktx:2.2.0'
-}
-
-release {
-    checkTasks = ['check']
-    buildTasks = ['publish']
-    updateVersionPart = 2
-    tagPrefix = 'android-ktx-'
-}
-
-publishing {
-    publications {
-        mavenAar(MavenPublication) {
-            from components.android
-        }
-    }
-}
