--- android-ktx/gradle.properties
@@ -1,2 +1,0 @@
-#Fri Dec 04 11:24:27 EST 2020
-version=1.0.1-SNAPSHOT
