--- android-ktx/src/main/AndroidManifest.xml
@@ -1,5 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.mirego.trikot.bluetooth.android.ktx">
-     <uses-permission android:name="android.permission.BLUETOOTH" />
-     <uses-permission android:name="android.permission.BLUETOOTH_ADMIN" />
-</manifest>
