--- bluetooth/build.gradle
@@ -1,89 +1,0 @@
-plugins {
-    id 'com.android.library'
-    id 'kotlin-multiplatform'
-    id 'org.jlleitschuh.gradle.ktlint'
-    id 'mirego.release' version '2.0'
-    id 'mirego.publish' version '1.0'
-}
-
-repositories {
-    google()
-    jcenter()
-    mavenLocal()
-    mavenCentral()
-    maven { url 'https://jitpack.io' }
-    maven { url('https://s3.amazonaws.com/mirego-maven/public') }
-}
-
-group 'com.mirego.trikot'
-
-android {
-    compileSdkVersion 30
-
-    defaultConfig {
-        minSdkVersion 14
-        targetSdkVersion 30
-    }
-}
-
-kotlin {
-    android {
-        publishAllLibraryVariants()
-    }
-
-    jvm()
-    ios()
-    iosArm32('iosArm32')
-    tvos()
-
-    sourceSets {
-        all {
-            languageSettings {
-                useExperimentalAnnotation('kotlin.time.ExperimentalTime')
-            }
-        }
-
-        commonMain {
-            dependencies {
-                implementation "com.mirego.trikot:streams:$trikot_streams_version"
-                implementation "com.mirego.trikot:trikotFoundation:$trikot_foundation_version"
-            }
-        }
-
-        androidMain {
-            dependsOn commonMain
-            dependencies {
-                implementation "com.mirego.trikot:android-ktx:$trikot_streams_android_ktx_version"
-            }
-        }
-
-        androidTest {
-            dependencies {
-                implementation 'org.jetbrains.kotlin:kotlin-test'
-                implementation 'org.jetbrains.kotlin:kotlin-test-junit'
-            }
-        }
-
-        jvmMain {
-            dependsOn commonMain
-        }
-
-        nativeMain {
-            dependsOn commonMain
-        }
-
-        iosArm64Main {
-            dependsOn nativeMain
-        }
-
-        iosX64Main {
-            dependsOn(nativeMain)
-        }
-    }
-}
-
-release {
-   checkTasks = ['check']
-   buildTasks = ['publish']
-   updateVersionPart = 2
-}
