+++ bluetooth/build.gradle.kts
@@ -1,0 +1,99 @@
+plugins {
+    kotlin("multiplatform")
+    id("com.android.library")
+    id("org.jlleitschuh.gradle.ktlint")
+    id("mirego.release").version("2.0")
+    id("mirego.publish").version("1.0")
+}
+
+repositories {
+    google()
+    mavenLocal()
+    mavenCentral()
+    maven("https://kotlin.bintray.com/kotlinx")
+    maven("https://jitpack.io")
+    maven("https://plugins.gradle.org/m2/")
+    maven("https://s3.amazonaws.com/mirego-maven/public")
+}
+
+group = "com.mirego.trikot"
+
+kotlin {
+    android {
+        publishAllLibraryVariants()
+    }
+    jvm()
+    ios()
+    iosArm32("iosArm32")
+    tvos()
+
+    sourceSets {
+        all {
+            languageSettings.useExperimentalAnnotation("kotlin.time.ExperimentalTime")
+        }
+
+        val commonMain by getting {
+            dependencies {
+                implementation("com.mirego.trikot:trikotFoundation:${project.extra["trikot_foundation_version"]}")
+                implementation("com.mirego.trikot:streams:${project.extra["trikot_streams_version"]}")
+            }
+        }
+
+        val jvmMain by getting {
+            dependsOn(commonMain)
+        }
+
+        val androidMain by getting {
+            dependsOn(commonMain)
+            dependencies {
+                implementation("androidx.lifecycle:lifecycle-extensions:2.2.0")
+                implementation("androidx.lifecycle:lifecycle-reactivestreams-ktx:2.3.1")
+            }
+        }
+
+        val androidTest by getting {
+            dependencies {
+                implementation("org.jetbrains.kotlin:kotlin-test")
+                implementation("org.jetbrains.kotlin:kotlin-test-junit")
+            }
+        }
+
+        val nativeMain by creating {
+            dependsOn(commonMain)
+        }
+
+        val iosArm32Main by getting {
+            dependsOn(nativeMain)
+        }
+
+        val iosArm64Main by getting {
+            dependsOn(nativeMain)
+        }
+
+        val iosX64Main by getting {
+            dependsOn(nativeMain)
+        }
+
+        val tvosArm64Main by getting {
+            dependsOn(nativeMain)
+        }
+
+        val tvosX64Main by getting {
+            dependsOn(nativeMain)
+        }
+    }
+}
+
+android {
+    defaultConfig {
+        compileSdkVersion(30)
+        minSdkVersion(21)
+        targetSdkVersion(30)
+    }
+}
+
+release {
+    checkTasks = listOf("check")
+    buildTasks = listOf("publish")
+    updateVersionPart = 2
+}
