+++ bluetooth/gradle.properties
@@ -1,2 +1,2 @@
 #Fri Dec 04 11:21:56 EST 2020
-version=1.0.2-SNAPSHOT
+version=2.0.0-SNAPSHOT
