+++ bluetooth/src/androidMain/kotlin/com/mirego/trikot/bluetooth/android/AndroidAttributeProfileService.kt
@@ -1,12 +1,11 @@
-package mirego.trikot.bluetooth
+package com.mirego.trikot.bluetooth.android
 
 import android.bluetooth.BluetoothGattService
 import com.mirego.trikot.bluetooth.AttributeProfileCharacteristic
 import com.mirego.trikot.bluetooth.AttributeProfileService
 import com.mirego.trikot.streams.reactive.Publishers
 import org.reactivestreams.Publisher
-import java.util.*
-import kotlin.collections.HashMap
+import java.util.Locale
 
 class AndroidAttributeProfileService(
     value: BluetoothGattService,
@@ -16,15 +15,20 @@ class AndroidAttributeProfileService(
         HashMap<String, AndroidAttributeProfileCharacteristic>()
             as Map<String, AndroidAttributeProfileCharacteristic>
     ) { result, value ->
-        (mutableMapOf(
-            value.uuid.toString().toUpperCase(Locale.ROOT) to AndroidAttributeProfileCharacteristic(
-                value,
-                bluetoothDevice
-            )
-        ) + result).toMap()
+        (
+            mutableMapOf(
+                value.uuid.toString()
+                    .toUpperCase(Locale.ROOT) to
+                    AndroidAttributeProfileCharacteristic(
+                        value,
+                        bluetoothDevice
+                    )
+            ) + result
+            ).toMap()
     }
 
     @Suppress("UNCHECKED_CAST")
-    override val characteristics = Publishers.behaviorSubject(androidCharacteristic
+    override val characteristics = Publishers.behaviorSubject(
+        androidCharacteristic
     ) as Publisher<Map<String, AttributeProfileCharacteristic>>
 }
