+++ bluetooth/src/main/AndroidManifest.xml
@@ -1,2 +1,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.mirego.trikot.bluetooth.android"/>
+    package="com.mirego.trikot.bluetooth.android">
+
+    <uses-permission android:name="android.permission.BLUETOOTH" />
+    <uses-permission android:name="android.permission.BLUETOOTH_ADMIN" />
+</manifest>
