--- build.gradle
@@ -1,26 +1,0 @@
-buildscript {
-    repositories {
-        google()
-        jcenter()
-        maven { url "https://plugins.gradle.org/m2/" }
-    }
-
-    dependencies {
-        classpath 'com.android.tools.build:gradle:3.5.0'
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-        classpath 'org.jlleitschuh.gradle:ktlint-gradle:8.0.0'
-    }
-}
-
-repositories {
-    google()
-    jcenter()
-}
-
-allprojects {
-    repositories {
-        google()
-        jcenter()
-        maven { url "https://kotlin.bintray.com/kotlinx" }
-    }
-}
