+++ gradle.properties
@@ -1,9 +1,9 @@
 kotlin.code.style=official
-kotlin_version=1.4.20
-trikot_foundation_version=1.0.21
-trikot_streams_version=1.0.6
-trikot_streams_android_ktx_version=1.0.1
+kotlin_version=1.4.30
+trikot_foundation_version=2.0.3
+trikot_streams_version=2.0.5
 kotlin.mpp.enableGranularSourceSetsMetadata=true
+kotlin.mpp.stability.nowarn=true
 kotlin.native.enableDependencyPropagation=false
 android.useAndroidX=true
 android.enableJetifier=true
