--- settings.gradle
@@ -1,26 +1,0 @@
-pluginManagement {
-    resolutionStrategy {
-        repositories {
-            mavenLocal()
-            maven {
-                url 'https://plugins.gradle.org/m2/'
-            }
-            maven {
-                url('https://s3.amazonaws.com/mirego-maven/public')
-            }
-            jcenter()
-            maven { url 'https://jitpack.io' }
-        }
-        eachPlugin {
-            if (requested.id.id in ['kotlin-multiplatform', 'kotlin-platform-native']) {
-                useModule("org.jetbrains.kotlin:${requested.id.id}:${requested.version}")
-            }
-            if (requested.id.namespace == 'mirego') {
-                useModule "mirego:${requested.id.name}-plugin:${requested.version}"
-            }
-        }
-    }
-}
-rootProject.name = 'trikot.bluetooth'
-
-include ':bluetooth', ':android-ktx'
